requirements.txt
aiogram>=3.13.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.27.0
groq>=0.4.0
